Move auth structs into lnurl/auth/model
