+++ libs/sdk-common/src/lib.rs
@@ -9,11 +9,12 @@ pub mod prelude {
     pub use crate::invoice::*;
     pub use crate::lnurl::error::*;
     pub use crate::lnurl::model::*;
+    pub use crate::lnurl::specs::auth::model::*;
     pub use crate::lnurl::specs::auth::*;
     pub use crate::lnurl::specs::pay::model::*;
     pub use crate::lnurl::specs::pay::*;
-    pub use crate::lnurl::specs::withdraw::*;
     pub use crate::lnurl::specs::withdraw::model::*;
+    pub use crate::lnurl::specs::withdraw::*;
     pub use crate::lnurl::*;
     pub use crate::model::*;
     pub use crate::utils::rest_client::*;
