--- libs/sdk-core/src/lnurl/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod withdraw;
